+++ lib/src/widgets/framework.dart
@@ -16,7 +16,7 @@ export 'package:flutter/foundation.dart' show FlutterError, ErrorSummary, ErrorD
 class $Widget implements Widget, $Instance {
   /// Compile-time type reference to [$Widget]
   static const $type =
-      BridgeTypeRef(BridgeTypeSpec('package:flutter/widgets.dart', 'Widget'));
+      BridgeTypeRef(BridgeTypeSpec('package:flutter/src/widgets/framework.dart', 'Widget'));
 
   /// Compile-time wrapper declaration for [$Widget]
   static const $declaration = BridgeClassDef(